+++ df-downloader-ui/src/routes/tools/tools.routes.tsx
@@ -1,7 +1,7 @@
 import HandymanIcon from "@mui/icons-material/Handyman";
 import DriveFileMove from "@mui/icons-material/DriveFileMove";
 import { NestedSubRoute } from "../nav/nested-routes.tsx";
-import { ReorgnaizeFilesPage } from "../../components/tools/batch-move-files/reorganize-files-page.component.tsx";
+import { ReorganizeFilesPage } from "../../components/tools/batch-move-files/reorganize-files-page.component.tsx";
 import { MaintenanceToolsPage } from "../../components/tools/maintenance/maintenance-tools-page.tsx";
 
 export const toolsRouteDefinitions: NestedSubRoute = {
@@ -10,8 +10,8 @@ export const toolsRouteDefinitions: NestedSubRoute = {
   routes: [
     {
       path: "/tools/reorganize-files",
-      element: <ReorgnaizeFilesPage />,
-      name: "Reorgnaize Files",
+      element: <ReorganizeFilesPage />,
+      name: "Reorganize Files",
       icon: DriveFileMove,
     },
     {